test(captp): split gc tests into separate files
