+++ packages/captp/test/test-finalizing-map.js
@@ -1,0 +1,27 @@
+import { test } from './prepare-test-env-ava.js';
+
+import { detectEngineGC } from './engine-gc.js';
+import { makeGcAndFinalize } from './gc-and-finalize.js';
+import { makeFinalizingMap } from '../src/finalize.js';
+
+const setAndDrop = async (t, map, droppedKey) => {
+  const obj = {};
+  map.set(droppedKey, obj);
+  t.is(map.get(droppedKey), obj);
+};
+
+test('finalizing map', async t => {
+  const gcAndFinalize = await makeGcAndFinalize(detectEngineGC());
+
+  const droppedKey = 'dropped';
+  const map = makeFinalizingMap(key => t.is(key, droppedKey));
+
+  const preserved = {};
+  map.set('preserved', preserved);
+  setAndDrop(t, map, droppedKey);
+
+  t.is(map.getSize(), 2);
+  await gcAndFinalize();
+  t.is(map.getSize(), 1);
+  t.is(map.get('preserved'), preserved);
+});
